requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
asyncpg==0.27.0
python-dotenv==1.0.0
alembic==1.11.1
pydantic==1.10.7